优化api状态改变处理

+++ crds/api/controllers/api_controller.go
@@ -81,10 +81,11 @@ func (r *ApiReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	}
 
 	//first publish   UnRelease--> publish 未发布-->执行发布api-->已发布
+	//TODO 下线时是否要删除route   每次发布都重新创建 若删除则API所有的关联都需要重新创建
 	if api.Status.Action == nlptv1.Publish && api.Status.PublishStatus == nlptv1.UnRelease {
 		// call kong api create
 		api.Status.Status = nlptv1.Running
-		klog.Infof("new api: %s:%d, cafile: %s", r.Operator.Host, r.Operator.Port, r.Operator.CAFile)
+		klog.Infof("new api: %s:%s, publish status: %s", api.Spec.Name, api.ObjectMeta.Name, api.Status.PublishStatus)
 		if err := r.Operator.CreateRouteByKong(api); err != nil {
 			api.Status.Status = nlptv1.Error
 			api.Status.Message = err.Error()
@@ -105,6 +106,7 @@ func (r *ApiReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 				r.Update(ctx, api)
 				return ctrl.Result{}, nil
 			}
+
 		}
 		if api.Spec.ApiDefineInfo.Cors == "true" {
 			if err := r.Operator.AddRouteCorsByKong(api); err != nil {
@@ -121,21 +123,26 @@ func (r *ApiReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 			r.Operator.Host, r.Operator.KongPortalPort, api.Spec.KongApi.Paths[0]))
 		api.Status.Message = "success"
 		r.Update(ctx, api)
+		return ctrl.Result{}, nil
 	}
 
 	//下线 api   已经发状态才可以执行下线   已发布-->执行下线api->已下线
 	if api.Status.Action == nlptv1.Offline && api.Status.PublishStatus == nlptv1.Released {
 		api.Status.Status = nlptv1.Running
 		// call kong api delete
-		if err := r.Operator.DeleteRouteByKong(api); err != nil {
+		if err := r.Operator.UpdateRouteInfoFromKong(api, true); err != nil {
 			//TODO 异常处理
 			api.Status.Status = nlptv1.Error
 			api.Status.Message = err.Error()
+			r.Update(ctx, api)
+			return ctrl.Result{}, nil
+
 		}
 		//PublishStatus -> offline
 		api.Status.Status = nlptv1.Success
 		api.Status.PublishStatus = nlptv1.Offlined
 		r.Update(ctx, api)
+		return ctrl.Result{}, nil
 	}
 
 	//发布后的API需要修改 需要先将API下线 之后更新 然后再执行发布  已下线-->执行更新然后发布-->已发布
@@ -146,22 +153,45 @@ func (r *ApiReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 			//TODO 异常处理
 			api.Status.Status = nlptv1.Error
 			api.Status.Message = err.Error()
+			r.Update(ctx, api)
+			return ctrl.Result{}, nil
 		}
 		//PublishStatus -> Released
 		api.Status.Status = nlptv1.Success
 		api.Status.PublishStatus = nlptv1.Released
 		r.Update(ctx, api)
+		return ctrl.Result{}, nil
 	}
-
-	if api.Status.Action == nlptv1.Delete {
+	//未发布API更新 已下线API更新
+	if api.Status.Action == nlptv1.Update && (api.Status.PublishStatus == nlptv1.UnRelease ||
+		api.Status.PublishStatus == nlptv1.Offlined) {
+		api.Status.Status = nlptv1.Success
+		klog.Infof("update api : %s %s, status %s", api.Spec.Name, api.ObjectMeta.Name, api.Status.PublishStatus)
+		r.Update(ctx, api)
+		return ctrl.Result{}, nil
+	}
+	//已下线删除 已下线时route已经删除
+	if api.Status.Action == nlptv1.Delete && api.Status.PublishStatus == nlptv1.Offlined {
 		api.Status.Status = nlptv1.Running
-		// call kong api delete
 		if err := r.Operator.DeleteRouteByKong(api); err != nil {
+			//TODO 异常处理
 			api.Status.Status = nlptv1.Error
 			api.Status.Message = err.Error()
+			r.Update(ctx, api)
+			return ctrl.Result{}, nil
+
 		}
 		api.Status.Status = nlptv1.Success
+		klog.Infof("delete api : %s %s, status", api.Spec.Name, api.ObjectMeta.Name, api.Status.PublishStatus)
 		r.Delete(ctx, api)
+		return ctrl.Result{}, nil
+	}
+	//未发布状态删除  未发布时还未创建route
+	if api.Status.Action == nlptv1.Delete && api.Status.PublishStatus == nlptv1.UnRelease {
+		api.Status.Status = nlptv1.Success
+		klog.Infof("delete api unreleased: %s %s", api.Spec.Name, api.ObjectMeta.Name)
+		r.Delete(ctx, api)
+		return ctrl.Result{}, nil
 	}
 
 	/*if api.Status.Status == nlptv1.Error {
